[CS-3313]: Create usePaymentLinks hook
